// plugins/nx-php/src/schemas/composer.schema.d.ts
export type Authors = {
    /**
     * Full name of the author.
     */
    name: string;
    /**
     * Email address of the author.
     */
    email?: string;
    /**
     * Homepage URL for the author.
     */
    homepage?: string;
    /**
     * Author's role in the project.
     */
    role?: string;
}[];
export type Repository =
    | ComposerRepository
    | VcsRepository
    | PathRepository
    | ArtifactRepository
    | PearRepository
    | PackageRepository;

export interface Package {
// ...
    'require-dev'?: {
        [k: string]: string;
    };
// ...
    'autoload-dev'?: {
        /**
         * This is an object of namespaces (keys) and the directories they can be found into (values, can be arrays of paths) by the autoloader.
         */
        'psr-0'?: {
            [k: string]: string | string[];
        };
        /**
         * This is an object of namespaces (keys) and the PSR-4 directories they can map to (values, can be arrays of paths) by the autoloader.
         */
        'psr-4'?: {
            [k: string]: string | string[];
        };
        /**
         * This is an array of paths that contain classes to be included in the class-map generation process.
         */
        classmap?: unknown[];
        /**
         * This is an array of files that are always required on every request.
         */
        files?: unknown[];
        [k: string]: unknown;
    };
// ...
    config?: {
// ...
        'vendor-dir'?: string;
// ...
    };
// ...
}
// ...
export interface Autoload {
    /**
     * This is an object of namespaces (keys) and the directories they can be found in (values, can be arrays of paths) by the autoloader.
     */
    'psr-0'?: {
        [k: string]: string | string[];
    };
    /**
     * This is an object of namespaces (keys) and the PSR-4 directories they can map to (values, can be arrays of paths) by the autoloader.
     */
    'psr-4'?: {
        [k: string]: string | string[];
    };
    /**
     * This is an array of paths that contain classes to be included in the class-map generation process.
     */
    classmap?: unknown[];
    /**
     * This is an array of files that are always required on every request.
     */
    files?: unknown[];
    /**
     * This is an array of patterns to exclude from autoload classmap generation. (e.g. "exclude-from-classmap": ["/test/", "/tests/", "/Tests/"]
     */
    'exclude-from-classmap'?: unknown[];
    [k: string]: unknown;
}